+++ modules/protocol/src/main/java/com/icodesoftware/protocol/GameServiceProvider.java
@@ -4,6 +4,6 @@ import com.icodesoftware.Room;
 
 public interface GameServiceProvider {
 
-    void updateStatistics(Room room, String system,int stub, String name, double delta);
-    void updateExperience(Room room,String system,int stub,double delta);
+    void updateStatistics(Room room, String system,long stub, String name, double delta);
+    void updateExperience(Room room,String system,long stub,double delta);
 }
